optimize: change VO classes name to Mcp
