+++ console/src/main/java/org/apache/seata/mcp/entity/vo/McpBranchSessionVO.java
@@ -21,31 +21,14 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import org.apache.seata.mcp.core.config.TimestampToStringDeserializer;
 
-import java.util.Set;
-
 import static com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility.ANY;
 import static com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility.NONE;
 
-/**
- * GlobalSessionVO
- */
 @JsonAutoDetect(getterVisibility = NONE, fieldVisibility = ANY)
-public class GlobalSessionVO extends org.apache.seata.server.console.entity.vo.GlobalSessionVO {
+public class McpBranchSessionVO extends org.apache.seata.server.console.entity.vo.BranchSessionVO {
 
-    private String beginTime;
     private String gmtCreate;
     private String gmtModified;
-    private Set<BranchSessionVO> branchSessionVOs;
-
-    @JsonProperty("beginTime")
-    @JsonDeserialize(using = TimestampToStringDeserializer.class)
-    public String getBegin() {
-        return beginTime;
-    }
-
-    public void setBeginTime(String beginTime) {
-        this.beginTime = beginTime;
-    }
 
     @JsonProperty("gmtCreate")
     @JsonDeserialize(using = TimestampToStringDeserializer.class)
@@ -65,14 +48,5 @@ public class GlobalSessionVO extends org.apache.seata.server.console.entity.vo.G
 
     public void setGmtModified(String gmtModified) {
         this.gmtModified = gmtModified;
-    }
-
-    @JsonProperty("branchSessionVOs")
-    public Set<BranchSessionVO> getBranchSessionVO() {
-        return branchSessionVOs;
-    }
-
-    public void setBranchSessionVO(Set<BranchSessionVO> branchSessionVOs) {
-        this.branchSessionVOs = branchSessionVOs;
     }
 }
